Consolidate overloaded setters in LuckyDrawModel for Firestore CustomClassMapper compatibility

+++ app/src/main/java/com/app/rewardsplanet/lucky_draw/activity_lucky_draw.java
@@ -698,7 +698,7 @@ public class activity_lucky_draw extends AppCompatActivity
                                                             m.getMyTicketsCount() + count
                                                     );
                                                 }
-                                                m.setFilledSlots(m.getFilledSlots() + count);
+                                                m.setFilledSlots((long) (m.getFilledSlots() + count));
                                                 break;
                                             }
                                         }
